requirements.txt
gradio
sentence-transformers
transformers
rank-bm25
faiss-cpu
nltk
numpy
torch
